Using the libgdx Animation class to render the movements of the player and flamer
Also removed some of the images to clean the assets folder. Instead of having to duplicate an image of the player/flamer facing left and flip it so it it facing right I now only have one image which I flip in code to face the other way. Note, there is only one flamer image, so using the Animation class is not necessary but I have placed the functionality there so it can be easily changed later if needed.